+++ template/about.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+{% block title %}About{% endblock %}
+{% block content %}
+{% include 'nav_panel.html' %}
+<h1>About</h1>
+<p>Somethink</p>
+{% endblock %}
