+++ template/index.html
@@ -1,0 +1,6 @@
+{% extends 'base.html' %}
+{% block title %}Main{% endblock %}
+{% block content %}
+{% include 'nav_panel.html' %}
+<h1>Main page</h1>
+{% endblock %}
