+++ template/nav_panel.html
@@ -1,0 +1,5 @@
+<ul>
+    <li><a href="{{ context['index'] }}">Main page</a></li>
+    <li><a href="{{ context['about'] }}">About</a></li>
+    <li><a href="{{ context['contacts'] }}">Contacts</a></li>
+</ul>
