+++ template/success_page.html
@@ -1,0 +1,6 @@
+{% extends 'base.html' %}
+{% block title %}Success{% endblock %}
+{% block content %}
+<p>Thanks for you message</p>
+<p><a href="/">Back to main page</a></p>
+{% endblock %}
