+++ urls.py
@@ -1,10 +1,11 @@
-from url_dispatcher import UrlTree, UrlNode
-from views import index_view, about_view
+from Framework.url_dispatcher import UrlTree, UrlNode
+from views import AboutView, IndexView, ContactsView, SuccessView
 
 front_controller = UrlTree()
-front_controller.root.controller = index_view
-about = UrlNode('about', about_view, front_controller.root)
+front_controller.root.controller = IndexView()
+about = UrlNode('about', AboutView(), front_controller.root)
+contacts = UrlNode('contacts', ContactsView(), front_controller.root)
+success = UrlNode('success', SuccessView(), front_controller.root)
 
-urls = [about]
+urls = [about, contacts, success]
 
-front_controller.create_tree(urls)
